+++ src/hotspot/share/opto/memnode.hpp
@@ -43,7 +43,7 @@ private:
   bool _unaligned_access; // Unaligned access from unsafe
   bool _mismatched_access; // Mismatched access from unsafe: byte read in integer array for instance
   bool _unsafe_access;     // Access of unsafe origin.
-  uint8_t _barrier_data;   // Bit field with barrier information
+  uint16_t _barrier_data;   // Bit field with barrier information
 
 protected:
 #ifdef ASSERT
@@ -140,9 +140,9 @@ public:
 #endif
   }
 
-  uint8_t barrier_data() { return _barrier_data; }
-  void set_barrier_data(uint8_t barrier_data) { _barrier_data = barrier_data; }
-  void add_barrier_data(uint8_t barrier_data) { _barrier_data |= barrier_data; }
+  uint16_t barrier_data() { return _barrier_data; }
+  void set_barrier_data(uint16_t barrier_data) { _barrier_data = barrier_data; }
+  void add_barrier_data(uint16_t barrier_data) { _barrier_data |= barrier_data; }
 
   // Search through memory states which precede this node (load or store).
   // Look for an exact match for the address, with no intervening
@@ -231,7 +231,7 @@ public:
                     const TypePtr* at, const Type *rt, BasicType bt,
                     MemOrd mo, ControlDependency control_dependency = DependsOnlyOnTest,
                     bool unaligned = false, bool mismatched = false, bool unsafe = false,
-                    uint8_t barrier_data = 0);
+                    uint16_t barrier_data = 0);
 
   virtual uint hash()   const;  // Check the type
 
@@ -417,7 +417,7 @@ public:
   bool require_atomic_access() const { return _require_atomic_access; }
   static LoadLNode* make_atomic(Node* ctl, Node* mem, Node* adr, const TypePtr* adr_type,
                                 const Type* rt, MemOrd mo, ControlDependency control_dependency = DependsOnlyOnTest,
-                                bool unaligned = false, bool mismatched = false, bool unsafe = false, uint8_t barrier_data = 0);
+                                bool unaligned = false, bool mismatched = false, bool unsafe = false, uint16_t barrier_data = 0);
 #ifndef PRODUCT
   virtual void dump_spec(outputStream *st) const {
     LoadNode::dump_spec(st);
@@ -469,7 +469,7 @@ public:
   bool require_atomic_access() const { return _require_atomic_access; }
   static LoadDNode* make_atomic(Node* ctl, Node* mem, Node* adr, const TypePtr* adr_type,
                                 const Type* rt, MemOrd mo, ControlDependency control_dependency = DependsOnlyOnTest,
-                                bool unaligned = false, bool mismatched = false, bool unsafe = false, uint8_t barrier_data = 0);
+                                bool unaligned = false, bool mismatched = false, bool unsafe = false, uint16_t barrier_data = 0);
 #ifndef PRODUCT
   virtual void dump_spec(outputStream *st) const {
     LoadNode::dump_spec(st);
@@ -839,7 +839,7 @@ class LoadStoreNode : public Node {
 private:
   const Type* const _type;      // What kind of value is loaded?
   const TypePtr* _adr_type;     // What kind of memory is being addressed?
-  uint8_t _barrier_data;        // Bit field with barrier information
+  uint16_t _barrier_data;        // Bit field with barrier information
   virtual uint size_of() const; // Size is bigger
 public:
   LoadStoreNode( Node *c, Node *mem, Node *adr, Node *val, const TypePtr* at, const Type* rt, uint required );
@@ -854,9 +854,9 @@ public:
   bool result_not_used() const;
   MemBarNode* trailing_membar() const;
 
-  uint8_t barrier_data() { return _barrier_data; }
-  void set_barrier_data(uint8_t barrier_data) { _barrier_data = barrier_data; }
-  void add_barrier_data(uint8_t barrier_data) { _barrier_data |= barrier_data; }
+  uint16_t barrier_data() { return _barrier_data; }
+  void set_barrier_data(uint16_t barrier_data) { _barrier_data = barrier_data; }
+  void add_barrier_data(uint16_t barrier_data) { _barrier_data |= barrier_data; }
 };
 
 class LoadStoreConditionalNode : public LoadStoreNode {
